+++ src/stat.cc
@@ -1,6 +1,6 @@
 
 /*
- * $Id: stat.cc,v 1.348 2001/08/16 00:16:18 hno Exp $
+ * $Id: stat.cc,v 1.349 2001/10/24 06:16:17 hno Exp $
  *
  * DEBUG: section 18    Cache Manager Statistics
  * AUTHOR: Harvest Derived
@@ -1409,8 +1409,8 @@ statClientRequests(StoreEntry * s)
 	    storeAppendPrintf(s, "\tFD %d, read %d, wrote %d\n", fd,
 		fd_table[fd].bytes_read, fd_table[fd].bytes_written);
 	    storeAppendPrintf(s, "\tFD desc: %s\n", fd_table[fd].desc);
-	    storeAppendPrintf(s, "\tin: buf %p, offset %d, size %d\n",
-		conn->in.buf, conn->in.offset, conn->in.size);
+	    storeAppendPrintf(s, "\tin: buf %p, offset %ld, size %ld\n",
+		conn->in.buf, (long int)conn->in.offset, (long int)conn->in.size);
 	    storeAppendPrintf(s, "\tpeer: %s:%d\n",
 		inet_ntoa(conn->peer.sin_addr),
 		ntohs(conn->peer.sin_port));
@@ -1419,20 +1419,21 @@ statClientRequests(StoreEntry * s)
 		ntohs(conn->me.sin_port));
 	    storeAppendPrintf(s, "\tnrequests: %d\n",
 		conn->nrequests);
-	    storeAppendPrintf(s, "\tdefer: n %d, until %d\n",
-		conn->defer.n, conn->defer.until);
+	    storeAppendPrintf(s, "\tdefer: n %d, until %ld\n",
+		conn->defer.n, (long int)conn->defer.until);
 	}
 	storeAppendPrintf(s, "uri %s\n", http->uri);
 	storeAppendPrintf(s, "log_type %s\n", log_tags[http->log_type]);
-	storeAppendPrintf(s, "out.offset %d, out.size %d\n",
-	    http->out.offset, http->out.size);
+	storeAppendPrintf(s, "out.offset %ld, out.size %lu\n",
+	    (long int)http->out.offset, (unsigned long int)http->out.size);
 	storeAppendPrintf(s, "req_sz %d\n", http->req_sz);
 	e = http->entry;
 	storeAppendPrintf(s, "entry %p/%s\n", e, e ? storeKeyText(e->hash.key) : "N/A");
 	e = http->old_entry;
 	storeAppendPrintf(s, "old_entry %p/%s\n", e, e ? storeKeyText(e->hash.key) : "N/A");
-	storeAppendPrintf(s, "start %d.%06d (%f seconds ago)\n", http->start.tv_sec,
-	    http->start.tv_usec,
+	storeAppendPrintf(s, "start %ld.%06d (%f seconds ago)\n",
+	    (long int)http->start.tv_sec,
+	    (int)http->start.tv_usec,
 	    tvSubDsec(http->start, current_time));
 	storeAppendPrintf(s, "\n");
     }
